refactor(api): extract shared JellyfinUserApiController base for GetCurrentUserId/GetJellyfinUsername, dropping the verbatim duplicate between the two controllers
